--- src/main/java/bank/client/exceptions/CpfAlreadyExistsException.java
@@ -1,12 +1,0 @@
-package bank.client.exceptions;
-
-import org.springframework.http.HttpStatus;
-
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.CONFLICT)
-public class CpfAlreadyExistsException extends RuntimeException {
-    public CpfAlreadyExistsException(String message) {
-        super(message);
-    }
-}
