--- src/main/java/bank/client/exceptions/handler/CpfAlreadyExistsExceptionHandler.java
@@ -1,21 +1,0 @@
-package bank.client.exceptions.handler;
-
-import bank.client.exceptions.CpfAlreadyExistsException;
-import bank.client.exceptions.details.CpfAlreadyExistsExceptionDetails;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class CpfAlreadyExistsExceptionHandler {
-
-    @ExceptionHandler(CpfAlreadyExistsException.class)
-    public ResponseEntity<?> handlerCpfAlreadyExistsException (CpfAlreadyExistsException cpfAlreadyExistsException) {
-        CpfAlreadyExistsExceptionDetails exceptionDetails = CpfAlreadyExistsExceptionDetails.cpfAlreadyExistDetailsBuilder.newBuilder()
-                .details(cpfAlreadyExistsException.getMessage())
-                .build();
-        return new ResponseEntity<>(exceptionDetails, HttpStatus.CONFLICT);
-    }
-
-}
