--- src/main/java/bank/client/response/ClientResponse.java
@@ -1,13 +1,0 @@
-package bank.client.response;
-public class ClientResponse {
-    private final String name;
-
-    public ClientResponse( String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
-
