+++ src/main/java/bank/customer/exceptions/ClientDoesntExistException.java
@@ -1,0 +1,11 @@
+package bank.customer.exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.NOT_FOUND)
+public class ClientDoesntExistException extends RuntimeException {
+    public ClientDoesntExistException(String message) {
+        super(message);
+    }
+}
