+++ src/main/java/bank/customer/exceptions/CpfAlreadyExistsException.java
@@ -1,0 +1,12 @@
+package bank.customer.exceptions;
+
+import org.springframework.http.HttpStatus;
+
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.CONFLICT)
+public class CpfAlreadyExistsException extends RuntimeException {
+    public CpfAlreadyExistsException(String message) {
+        super(message);
+    }
+}
