+++ src/main/java/bank/customer/response/ClientResponse.java
@@ -1,0 +1,13 @@
+package bank.customer.response;
+public class ClientResponse {
+    private final String name;
+
+    public ClientResponse( String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
+
