+++ src/main/java/bank/customer/service/ClientService.java
@@ -1,0 +1,13 @@
+package bank.customer.service;
+
+import bank.model.Customer;
+
+import java.util.List;
+
+public interface ClientService {
+    List<Customer> getAllClients();
+    Customer getClientCpf(String cpf);
+    Long getClientId(String cpf);
+
+}
+
