+++ src/main/java/bank/transaction/service/TransactionServiceImpl.java
@@ -3,8 +3,8 @@ package bank.transaction.service;
 import bank.account.exceptions.AccountAlreadyExistsException;
 import bank.account.repository.AccountRepository;
 import bank.account.service.AccountServiceImpl;
-import bank.client.exceptions.ClientDoesntExistException;
-import bank.client.service.ClientServiceImpl;
+import bank.customer.exceptions.ClientDoesntExistException;
+import bank.customer.service.ClientServiceImpl;
 import bank.model.Account;
 import bank.model.Transaction;
 import bank.transaction.exception.ValueNotAcceptedException;
