+++ src/main/resources/db/migration/V002__account.sql
@@ -1,11 +1,11 @@
 create table account (
 id BIGSERIAL not null primary key,
 account_number bigint not null,
-client_id BIGSERIAL not null,
+customer_id BIGSERIAL not null,
 balance_money numeric not null,
 created_data timestamp,
 
-CONSTRAINT fk_client_id
-FOREIGN KEY (client_id)
-REFERENCES client(id)
+CONSTRAINT fk_customer_id
+FOREIGN KEY (customer_id)
+REFERENCES customer(id)
 );
