+++ src/test/java/bank/account/ServiceTests/AccountServiceImplTests.java
@@ -9,9 +9,9 @@ import bank.account.request.AccountRequest;
 import bank.account.service.AccountService;
 import bank.account.service.AccountServiceImpl;
 import bank.model.Account;
-import bank.model.Client;
-import bank.client.repository.ClientRepository;
-import bank.client.request.ClientRequest;
+import bank.model.Customer;
+import bank.customer.repository.ClientRepository;
+import bank.customer.request.ClientRequest;
 import bank.transaction.repository.TransactionRepository;
 import bank.transaction.service.TransactionServiceImpl;
 import org.junit.Before;
@@ -60,8 +60,8 @@ public class AccountServiceImplTests {
     @Spy
     ClientRequest clientRequest;
     ClientRequest clientRequest2;
-    Client client;
-    Client client2;
+    Customer customer;
+    Customer customer2;
 
     @Before
     public void init() {
@@ -74,23 +74,23 @@ public class AccountServiceImplTests {
         clientRequest = new ClientRequest("Victoria", "12345678901", "02036020", "SE", "SP", "SP");
         clientRequest2 = new ClientRequest("Thais", "12345678902", "02036020", "SE", "SP", "SP");
 
-        client = new Client();
-        client.setName("Victoria");
-        client.setCpf("12345678901");
-        client.setCity("SP");
-        client.setState("SP");
-        client.setStreet("SE");
-        client.setPostalCode("02036020");
-        clientRepository.save(client);
+        customer = new Customer();
+        customer.setName("Victoria");
+        customer.setCpf("12345678901");
+        customer.setCity("SP");
+        customer.setState("SP");
+        customer.setStreet("SE");
+        customer.setPostalCode("02036020");
+        clientRepository.save(customer);
 
-        client2 = new Client();
-        client2.setName("Thais");
-        client2.setCpf("12345678902");
-        client2.setCity("SP");
-        client2.setState("SP");
-        client2.setStreet("SE");
-        client2.setPostalCode("02036020");
-        clientRepository.save(client2);
+        customer2 = new Customer();
+        customer2.setName("Thais");
+        customer2.setCpf("12345678902");
+        customer2.setCity("SP");
+        customer2.setState("SP");
+        customer2.setStreet("SE");
+        customer2.setPostalCode("02036020");
+        clientRepository.save(customer2);
     }
 
     @Test
@@ -98,22 +98,22 @@ public class AccountServiceImplTests {
 
         Long accountNumber = accountRepository.generateAccountNumber();
         // Cria um cliente de teste
-        Client client = new Client();
-        client.setCpf("12345678901");
-        clientRepository.save(client);
+        Customer customer = new Customer();
+        customer.setCpf("12345678901");
+        clientRepository.save(customer);
 
         AccountRequest accountRequest = new AccountRequest();
         // Simula a busca pelo cliente recém-criado
-        when(clientRepository.findByCpf("12345678901")).thenReturn(Optional.of(client));
+        when(clientRepository.findByCpf("12345678901")).thenReturn(Optional.of(customer));
 
         // Registra uma conta para o cliente
         Account account = new Account();
-        account.setClient(client);
+        account.setCustomer(customer);
         account.setAccountNumber(accountNumber);
         account.setBalanceMoney(accountRequest.getBalanceMoney());
         // Verifica se a conta foi registrada corretamente
         assertNotNull(account);
-        assertEquals("12345678901", account.getClient().getCpf());
+        assertEquals("12345678901", account.getCustomer().getCpf());
     }
     @Test
     public void shouldThrowAccountAlreadyExistsException_whenRegisterAccount() {
@@ -121,8 +121,8 @@ public class AccountServiceImplTests {
         String cpf = "12345678901";
         Account account = new Account();
         account.setAccountNumber(1L);
-        account.setClient(new Client());
-        when(clientRepository.findByCpf(cpf)).thenReturn(Optional.of(new Client()));
+        account.setCustomer(new Customer());
+        when(clientRepository.findByCpf(cpf)).thenReturn(Optional.of(new Customer()));
         when(accountRepository.generateAccountNumber()).thenReturn(account.getAccountNumber());
         when(accountRepository.existsByAccountNumber(account.getAccountNumber())).thenReturn(true);
 
@@ -142,8 +142,8 @@ public class AccountServiceImplTests {
         String cpf = "12345678909";
         Account account = new Account();
         account.setAccountNumber(1L);
-        account.setClient(new Client());
-        when(clientRepository.findByCpf(cpf)).thenReturn(Optional.of(new Client()));
+        account.setCustomer(new Customer());
+        when(clientRepository.findByCpf(cpf)).thenReturn(Optional.of(new Customer()));
         when(accountRepository.generateAccountNumber()).thenReturn(account.getAccountNumber());
         when(accountRepository.existsByAccountNumber(account.getAccountNumber())).thenReturn(false);
         when(clientRepository.existsByCpf(cpf)).thenReturn(null);
@@ -214,15 +214,15 @@ public class AccountServiceImplTests {
         Long clientId = 1L;
         Long accountNumber = 123456L;
 
-        // Cria um objeto Client com uma conta associada
-        Client client = new Client();
-        client.setId(clientId);
+        // Cria um objeto Customer com uma conta associada
+        Customer customer = new Customer();
+        customer.setId(clientId);
         Account account = new Account();
         account.setAccountNumber(accountNumber);
-        client.setAccount(account);
+        customer.setAccount(account);
 
         // Define o comportamento do mock do repository
-        when(clientRepository.findById(clientId)).thenReturn(Optional.of(client));
+        when(clientRepository.findById(clientId)).thenReturn(Optional.of(customer));
 
         // Chama o método que deve retornar o número da conta
         Long result = accountServiceImpl.findAccountNumberByClientId(clientId);
@@ -235,16 +235,16 @@ public class AccountServiceImplTests {
     public void testFindAccountNumberByClientId_ThrowAccountDoesntExistException() {
         Long clientId = 1L;
 
-        Client client = new Client();
-        client.setId(clientId);
+        Customer customer = new Customer();
+        customer.setId(clientId);
 
         Account account = new Account();
         account.setAccountNumber(null);
-        account.setClient(client);
+        account.setCustomer(customer);
 
-        client.setAccount(account);
+        customer.setAccount(account);
 
-        Mockito.when(clientRepository.findById(clientId)).thenReturn(Optional.of(client));
+        Mockito.when(clientRepository.findById(clientId)).thenReturn(Optional.of(customer));
 
         try {
             accountServiceImpl.findAccountNumberByClientId(clientId);
@@ -256,12 +256,12 @@ public class AccountServiceImplTests {
     @Test
     public void testGetAccountById_ReturnsAccount() {
         Account account = new Account();
-        Client client = new Client();
-        client.setId(1L);
-        client.setAccount(account);
-        account.setClient(client);
+        Customer customer = new Customer();
+        customer.setId(1L);
+        customer.setAccount(account);
+        account.setCustomer(customer);
 
-        Mockito.when(clientRepository.findById(1L)).thenReturn(Optional.of(client));
+        Mockito.when(clientRepository.findById(1L)).thenReturn(Optional.of(customer));
 
         Optional<Account> returnedAccount = accountServiceImpl.getAccountById(1L);
 
@@ -271,7 +271,7 @@ public class AccountServiceImplTests {
     public void testGetAccountById_ThrowAccountDoesntExistException() {
         // given
         Long clientId = 1L;
-        Mockito.when(clientRepository.findById(clientId)).thenReturn(Optional.of(new Client()));
+        Mockito.when(clientRepository.findById(clientId)).thenReturn(Optional.of(new Customer()));
 
         try {
             // when
